Isolate execution of functional test related to tx pool (unconfirmed phases) - Reserves node-05 to the purpose - Moving unconfirmed phases for types 1 & 4 from node-02 to the new target
