+++ package.json
@@ -18,6 +18,7 @@
     "test-functional-http-get": "grunt test-functional-http-get",
     "test-functional-http-get-extensive": "grunt test-functional-http-get-extensive",
     "test-functional-http-post": "grunt test-functional-http-post",
+    "test-functional-pool": "grunt test-functional-pool",
     "jsdoc": "jsdoc -c docs/conf.json --verbose --pedantic",
     "create-bundles": "webpack",
     "server-docs": "npm run jsdoc && http-server docs/jsdoc/",
