+++ test/functional/pool/4.X.unconfirmed/4.0.transfer.js
@@ -1,27 +1,27 @@
 'use strict';
 
-var node = require('../../../../node');
-var shared = require('../../../shared');
+var node = require('../../../node');
+var shared = require('../../shared');
 var localShared = require('./shared');
 
-var sendTransactionPromise = require('../../../../common/apiHelpers').sendTransactionPromise;
+var sendTransactionPromise = require('../../../common/apiHelpers').sendTransactionPromise;
 
-describe('POST /api/transactions (unconfirmed type 6 on top of type 4)', function () {
+describe('POST /api/transactions (unconfirmed type 0 on top of type 4)', function () {
 
 	var scenarios = {
-		'regular': new shared.MultisigScenario(),
+		'regular': new shared.MultisigScenario()
 	};
 
 	var transaction;
 	var badTransactions = [];
 	var goodTransactions = [];
 
-	localShared.beforeValidationPhaseWithDapp(scenarios);
+	localShared.beforeValidationPhase(scenarios);
 
-	describe('sending inTransfer', function () {
+	describe('sending funds', function () {
 
 		it('regular scenario should be ok', function () {
-			transaction = node.lisk.transfer.createInTransfer(scenarios.regular.dapp.id, 1, scenarios.regular.account.password);
+			transaction = node.lisk.transaction.createTransaction(node.randomAccount().address, 1, scenarios.regular.account.password);
 
 			return sendTransactionPromise(transaction).then(function (res) {
 				node.expect(res).to.have.property('status').to.equal(200);
