+++ test/functional/pool/4.X.unconfirmed/4.2.delegate.js
@@ -1,12 +1,12 @@
 'use strict';
 
-var node = require('../../../../node');
-var shared = require('../../../shared');
+var node = require('../../../node');
+var shared = require('../../shared');
 var localShared = require('./shared');
 
-var sendTransactionPromise = require('../../../../common/apiHelpers').sendTransactionPromise;
+var sendTransactionPromise = require('../../../common/apiHelpers').sendTransactionPromise;
 
-describe('POST /api/transactions (unconfirmed type 3 on top of type 4)', function () {
+describe('POST /api/transactions (unconfirmed type 2 on top of type 4)', function () {
 
 	var scenarios = {
 		'regular': new shared.MultisigScenario(),
@@ -18,10 +18,10 @@ describe('POST /api/transactions (unconfirmed type 3 on top of type 4)', functio
 
 	localShared.beforeValidationPhase(scenarios);
 
-	describe('voting delegate', function () {
+	describe('registering delegate', function () {
 
 		it('regular scenario should be ok', function () {
-			transaction = node.lisk.vote.createVote(scenarios.regular.account.password, ['+' + node.eAccount.publicKey]);
+			transaction = node.lisk.delegate.createDelegate(scenarios.regular.account.password, scenarios.regular.account.username);
 
 			return sendTransactionPromise(transaction).then(function (res) {
 				node.expect(res).to.have.property('status').to.equal(200);
