+++ test/functional/pool/4.X.unconfirmed/4.3.votes.js
@@ -1,15 +1,15 @@
 'use strict';
 
-var node = require('../../../../node');
-var shared = require('../../../shared');
+var node = require('../../../node');
+var shared = require('../../shared');
 var localShared = require('./shared');
 
-var sendTransactionPromise = require('../../../../common/apiHelpers').sendTransactionPromise;
+var sendTransactionPromise = require('../../../common/apiHelpers').sendTransactionPromise;
 
-describe('POST /api/transactions (unconfirmed type 0 on top of type 4)', function () {
+describe('POST /api/transactions (unconfirmed type 3 on top of type 4)', function () {
 
 	var scenarios = {
-		'regular': new shared.MultisigScenario()
+		'regular': new shared.MultisigScenario(),
 	};
 
 	var transaction;
@@ -18,10 +18,10 @@ describe('POST /api/transactions (unconfirmed type 0 on top of type 4)', functio
 
 	localShared.beforeValidationPhase(scenarios);
 
-	describe('sending funds', function () {
+	describe('voting delegate', function () {
 
 		it('regular scenario should be ok', function () {
-			transaction = node.lisk.transaction.createTransaction(node.randomAccount().address, 1, scenarios.regular.account.password);
+			transaction = node.lisk.vote.createVote(scenarios.regular.account.password, ['+' + node.eAccount.publicKey]);
 
 			return sendTransactionPromise(transaction).then(function (res) {
 				node.expect(res).to.have.property('status').to.equal(200);
