+++ test/functional/pool/4.X.unconfirmed/4.5.dapps.js
@@ -1,12 +1,12 @@
 'use strict';
 
-var node = require('../../../../node');
-var shared = require('../../../shared');
+var node = require('../../../node');
+var shared = require('../../shared');
 var localShared = require('./shared');
 
-var sendTransactionPromise = require('../../../../common/apiHelpers').sendTransactionPromise;
+var sendTransactionPromise = require('../../../common/apiHelpers').sendTransactionPromise;
 
-describe('POST /api/transactions (unconfirmed type 1 on top of type 4)', function () {
+describe('POST /api/transactions (unconfirmed type 5 on top of type 4)', function () {
 
 	var scenarios = {
 		'regular': new shared.MultisigScenario(),
@@ -18,10 +18,10 @@ describe('POST /api/transactions (unconfirmed type 1 on top of type 4)', functio
 
 	localShared.beforeValidationPhase(scenarios);
 
-	describe('registering second password', function () {
+	describe('registering dapp', function () {
 
 		it('regular scenario should be ok', function () {
-			transaction = node.lisk.signature.createSignature(scenarios.regular.account.password, scenarios.regular.account.secondPassword);
+			transaction = node.lisk.dapp.createDapp(scenarios.regular.account.password, null, node.guestbookDapp);
 
 			return sendTransactionPromise(transaction).then(function (res) {
 				node.expect(res).to.have.property('status').to.equal(200);
