+++ test/functional/pool/4.X.unconfirmed/shared.js
@@ -1,14 +1,14 @@
 'use strict';
 
-var node = require('../../../../node');
-var shared = require('../../../shared');
+var node = require('../../../node');
+var shared = require('../../shared');
 
-var sendTransactionPromise = require('../../../../common/apiHelpers').sendTransactionPromise;
-var waitForConfirmations = require('../../../../common/apiHelpers').waitForConfirmations;
-var apiCodes = require('../../../../../helpers/apiCodes');
+var sendTransactionPromise = require('../../../common/apiHelpers').sendTransactionPromise;
+var waitForConfirmations = require('../../../common/apiHelpers').waitForConfirmations;
+var apiCodes = require('../../../../helpers/apiCodes');
 
-var sendTransactionPromise = require('../../../../common/apiHelpers').sendTransactionPromise;
-var sendSignaturePromise = require('../../../../common/apiHelpers').sendSignaturePromise;
+var sendTransactionPromise = require('../../../common/apiHelpers').sendTransactionPromise;
+var sendSignaturePromise = require('../../../common/apiHelpers').sendSignaturePromise;
 
 function beforeValidationPhase (scenarios) {
 	var transactionsToWaitFor = [];
